backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "uvicorn==0.34.2",
    "python-dotenv==1.1.0",
    "langchain==0.3.25",
    "langchain-openai==0.3.17",
    "langgraph==0.4.5",
    "langsmith==0.3.42",
    "beautifulsoup4",
    "playwright",
    "sse-starlette==2.1.3",
    "requests==2.32.3",
    "psycopg-pool",
    "langgraph-checkpoint-postgres",
    "tiktoken==0.9.0",
    "openai==1.79.0",
]
